# Fix replace command performance in Ex mode