--- a/changelog.d/20231023_220739_GitHub_Actions_fix-replace-performance.rst
+++ b/changelog.d/20231023_220739_GitHub_Actions_fix-replace-performance.rst
@@ -1,0 +1,7 @@
+.. _#1900:  https://github.com/fox0430/moe/pull/1900
+
+Fixed
+.....
+
+- `#1900`_ Fix replace command performance in Ex mode
+