Fix the replace command performance in Ex mode. (`%s/xxx/yyy/`)

Speed up over about 1500x.

Executed `%s/proc/abc/` command in  `src/moepkg/editorstatus.nim`.
|Before|After|
|---|---|
|9 seconds, 365 milliseconds|5 milliseconds|